requirements.txt
flask
pdfminer.six
python-docx
sentence-transformers
numpy
torch
pymupdf
pytesseract
Pillow
